% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_straits_palette.R
\name{dire_straits_palette}
\alias{dire_straits_palette}
\title{Dire Straits Palette}
\usage{
dire_straits_palette(palette = "money_for_nothing")
}
\arguments{
\item{palette}{Select a color palette from dire straits album covers}
}
\value{
dire_straits_palette
}
\description{
Create a color palette from Dire Straits album covers
}
