% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direstraits_palette_gen.R
\name{direstraits_palette_gen}
\alias{direstraits_palette_gen}
\title{Generate the palette}
\usage{
direstraits_palette_gen(palette = "money_for_nothing", direction = 1, ...)
}
\arguments{
\item{palette}{Select a color palette from dire straits album covers}

\item{direction}{Select the direction of the color order}

\item{...}{More arguments}
}
\value{
direstraits_palette_gen
}
\description{
Create a color palette from Dire Straits album covers
}
