% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direstraits_color_scale.R
\name{scale_color_direstraits}
\alias{scale_color_direstraits}
\title{Create ggplot scale color direstraits}
\usage{
scale_color_direstraits(palette = "money_for_nothing", direction = 1, ...)
}
\arguments{
\item{palette}{Select a color palette from dire straits album covers}

\item{direction}{Select the direction of the color order}

\item{...}{More arguments}
}
\value{
scale_color_direstraits
}
\description{
Create a color palette from Dire Straits album covers
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(x = Sepal.Width, y = Sepal.Length, color = Species)) +
  ggplot2::geom_point(size = 5.0) +
  scale_color_direstraits("so_far_away") +
  ggplot2::theme_minimal()
}
