% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direstraits_fill_scale.R
\name{scale_fill_direstraits}
\alias{scale_fill_direstraits}
\title{Create ggplot scale fill for direstraits}
\usage{
scale_fill_direstraits(palette = "money_for_nothing", direction = 1, ...)
}
\arguments{
\item{palette}{Select a color palette from dire straits album covers}

\item{direction}{Select the direction of the color order}

\item{...}{More arguments}
}
\value{
scale_fill_direstraits
}
\description{
Create a color palette from Dire Straits album covers
}
