% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direstraits_color_scale.R
\name{scale_color_direstraits}
\alias{scale_color_direstraits}
\alias{scale_colour_direstraits}
\alias{scale_fill_direstraits}
\title{Create ggplot scale color Dire Straits}
\usage{
scale_colour_direstraits(palette = "money_for_nothing", direction = 1, ...)

scale_color_direstraits(palette = "money_for_nothing", direction = 1, ...)

scale_fill_direstraits(palette = "money_for_nothing", direction = 1, ...)
}
\arguments{
\item{palette}{Select a color palette from dire straits album covers}

\item{direction}{Select the direction of the color order}

\item{...}{More arguments}
}
\value{
scale_colour_direstraits

scale_color_direstraits

scale_fill_direstraits
}
\description{
Create ggplot scale color for Dire Straits.
}
\details{
colour palette from dire straits

color palette from dire straits

fill palette from dire straits
}
